+++ source/config/JsonScanner.ts
@@ -1,42 +1,5 @@
 import { DiagnosticOrigin, type SourceFile } from "#diagnostic";
-import type { OptionValue } from "./OptionDefinitionsMap.js";
-
-export class JsonElement {
-  origin: DiagnosticOrigin;
-  text: string | undefined;
-  value: OptionValue;
-
-  constructor(text: string | undefined, origin: DiagnosticOrigin) {
-    this.origin = origin;
-    this.text = text;
-
-    this.value = this.#resolveValue();
-  }
-
-  #resolveValue() {
-    if (this.text == null) {
-      return undefined;
-    }
-
-    if (/^['"]/.test(this.text)) {
-      return this.text.slice(1, -1);
-    }
-
-    if (this.text === "true") {
-      return true;
-    }
-
-    if (this.text === "false") {
-      return false;
-    }
-
-    if (/^\d/.test(this.text)) {
-      return Number.parseFloat(this.text);
-    }
-
-    return this.text;
-  }
-}
+import { JsonElement } from "./JsonElement.js";
 
 export class JsonScanner {
   #currentPosition = 0;
