+++ source/config/OptionDefinitionsMap.ts
@@ -8,8 +8,6 @@ export interface ItemDefinition {
 
 export type OptionDefinition = PrimitiveTypeOptionDefinition | ListTypeOptionDefinition;
 
-export type OptionValue = Array<OptionValue> | string | number | boolean | null | undefined;
-
 interface BaseOptionDefinition {
   brand: OptionBrand;
   description: string;
